requirements.txt
APScheduler==3.7.0
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
decorator==5.0.9
Flask==2.0.1
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pyTelegramBotAPI==3.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
aiogram~=2.14.3


ado-py==0.0.7
aiodns==2.0.0
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
cchardet==2.1.7
cffi==1.14.4
multidict==5.0.2
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
toml==0.10.2
typing-extensions==3.7.4.3
ujson==4.0.1
uvloop==0.14.0
yarl==1.6.3